podcast/requirements.txt
yt-dlp
boto3
python-dotenv
requests
flask
gunicorn
thefuzz[speed]
google-genai
